HotFix for login
As I realised the login didn't work quite as expected because xamarin called the OnStart later than the OnAppear method of the other ViewPages classes I made some adjustments. Now on app start a LoadingPage is loaded and than the OnStart event is fired logging the user in. This way we can be sure the user is logged in before accessing another page.